refact: extrair listeners de proficiency/enhancement para módulos dedicados (actor-sheet.mjs)
